+++ app/src/routes/api/v1/subscription.router.js
@@ -32,6 +32,14 @@ const serializeObjToQuery = (obj) => Object.keys(obj).reduce((a, k) => {
     return a;
 }, []).join('&');
 
+const getHostForPaginationLink = (ctx) => {
+    if ('referer' in ctx.request.header) {
+        const url = new URL(ctx.request.header.referer);
+        return url.host;
+    }
+    return ctx.request.host;
+};
+
 class SubscriptionsRouter {
 
     static getUser(ctx) {
@@ -300,7 +308,7 @@ class SubscriptionsRouter {
         delete clonedQuery.loggedUser;
         const serializedQuery = serializeObjToQuery(clonedQuery) ? `?${serializeObjToQuery(clonedQuery)}&` : '?';
         const apiVersion = ctx.mountPath.split('/')[ctx.mountPath.split('/').length - 1];
-        const link = `${ctx.request.protocol}://${ctx.request.host}/${apiVersion}${ctx.request.path}${serializedQuery}`;
+        const link = `${ctx.request.protocol}://${getHostForPaginationLink(ctx)}/${apiVersion}${ctx.request.path}${serializedQuery}`;
 
         const subscriptions = await SubscriptionService.getAllSubscriptions(
             link,